requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
argilla==0.0.1
attrs==23.1.0
backoff==2.2.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chromadb==0.3.26
click==8.1.7
clickhouse-connect==0.6.21
coloredlogs==15.0.1
cryptography==41.0.5
dataclasses-json==0.5.14
duckdb==0.9.2
et-xmlfile==1.1.0
fastapi==0.99.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
h11==0.14.0
hnswlib==0.7.0
httptools==0.6.1
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.221
langchainplus-sdk==0.0.20
lxml==4.9.3
lz4==4.3.2
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
olefile==0.46
onnxruntime==1.16.3
openai==0.28.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.1.0
posthog==3.0.2
protobuf==4.25.1
pulsar-client==3.3.0
pycparser==2.21
pydantic==1.10.13
pypandoc==1.12
pypdf==3.17.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.7.12
urllib3==2.1.0
uvicorn==0.22.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.3
zstandard==0.22.0
streamlit==1.25.0
PyPDF2==3.0.0
langchain==0.0.267
langchain-community==0.0.11
openai==0.30.0
scikit-learn==1.3.1
python-dotenv==1.0.0
numpy==1.24.4
chromadb==0.4.3
tiktoken==0.4.0
